# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Team:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      organization_id:
        type: string
      updated_at:
        type: string
    type: object
...
  domain.UserWithTeamRole:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      joined_at:
        type: string
      notification_preferences:
        additionalProperties: true
        type: object
      phone:
        type: string
      role:
        type: string
      timezone:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  service.AddTeamMemberRequest:
    properties:
      role:
        type: string
      user_id:
        type: string
    required:
    - user_id
    type: object
...
  service.CreateTeamRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
  service.TeamWithMembers:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      members:
        items:
          $ref: '#/definitions/domain.UserWithTeamRole'
        type: array
      name:
        type: string
      organization_id:
        type: string
      updated_at:
        type: string
    type: object
...
  service.UpdateTeamMemberRoleRequest:
    properties:
      role:
        type: string
    required:
    - role
    type: object
  service.UpdateTeamRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@pulsar.io
    name: API Support
    url: https://github.com/nmn3m/pulsar
  description: Pulsar is an incident management and alerting platform. This API provides
    endpoints for managing alerts, incidents, teams, schedules, escalation policies,
    notifications, and webhooks.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Pulsar API
  version: "1.0"
paths:
...
  /teams:
    get:
      consumes:
      - application/json
      description: List all teams in the organization
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/domain.Team'
              type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List teams
      tags:
      - Teams
    post:
      consumes:
      - application/json
      description: Create a new team in the organization
      parameters:
      - description: Create team request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateTeamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Team'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new team
      tags:
      - Teams
  /teams/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a team by ID
      parameters:
      - description: Team ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a team
      tags:
      - Teams
    get:
      consumes:
      - application/json
      description: Get a team by ID with its members
      parameters:
      - description: Team ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.TeamWithMembers'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a team
      tags:
      - Teams
    patch:
      consumes:
      - application/json
      description: Update a team by ID
      parameters:
      - description: Team ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Update team request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateTeamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Team'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a team
      tags:
      - Teams
  /teams/{id}/members:
    get:
      consumes:
      - application/json
      description: List all members of a team
      parameters:
      - description: Team ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/domain.UserWithTeamRole'
              type: array
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List team members
      tags:
      - Teams
    post:
      consumes:
      - application/json
      description: Add a user as a member to a team
      parameters:
      - description: Team ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Add member request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.AddTeamMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add a member to a team
      tags:
      - Teams
  /teams/{id}/members/{userId}:
    delete:
      consumes:
      - application/json
      description: Remove a user from a team
      parameters:
      - description: Team ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        format: uuid
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove a member from a team
      tags:
      - Teams
    patch:
      consumes:
      - application/json
      description: Update the role of a team member
      parameters:
      - description: Team ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        format: uuid
        in: path
        name: userId
        required: true
        type: string
      - description: Update role request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateTeamMemberRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a team member's role
      tags:
      - Teams
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"